[REF] mrp_brewing: remove parent_brew_order_id
parent_brew_order_id and child_brew_orders are never assigned.
On production database (renard), only 1/225 brew order has its value
assigned, it is an old order of 2017.
It's value is tested once in an incomplete condition test (cf
action_confirm). For that particular order, the test is not checked.

=> I decided to remove it for cleanliness and clarity.