[FIX] mrp_brewing: computing lot_id on sale order
On sale.order.line several stock.move may be found. Checking
`moves.mov_line_ids` failed because of `moves` not being a singleton.

Fixing that using mapped. Also this proposition is not totally the same.
Before the field `product_lot_ids` is updated only if there is only one
move. Now the field is always updated but with empty value when there is
none to set.